feat(freertos): Using align data to avoid exception handle

+++ components/freertos/port/esp8266/port.c
@@ -49,8 +49,8 @@
 #define PORT_ASSERT(x) do { if (!(x)) {ets_printf("%s %u\n", "rtos_port", __LINE__); while(1){}; }} while (0)
 
 extern char NMIIrqIsOn;
-static uint8_t HdlMacSig = 0;
-static uint8_t SWReq = 0;
+static uint32_t HdlMacSig = 0;
+static uint32_t SWReq = 0;
 
 unsigned cpu_sr;
